geotiff tests: consolidate GPU kernels + kwargs cluster (Sub-PR D)
Closes part of #2438 (Sub-PR D of four). Cluster 14 of the long-tail
GeoTIFF test consolidation epic (#2424). Tests-only.

Folds twelve top-level GPU kernel + kwarg-surface files into a new
xrspatial/geotiff/tests/gpu/test_kernels_and_kwargs.py, one Section
per source issue:

- test_gpu_cuda_preflight_1903.py
- test_gpu_kwarg_rename_1560.py
- test_gpu_strict_fallback_1516.py
- test_gpu_fallback_forwards_kwargs_2238.py
- test_kvikio_batched_pread_1688.py
- test_gds_chunked_gpu_parity_1896.py
- test_gds_fallback_batched_d2h_1552.py
- test_orientation_gpu.py
- test_size_param_validation_gpu_vrt_1776.py
- test_mask_nodata_gpu_vrt_2052.py
- test_open_geotiff_on_gpu_failure_1615.py
- test_crs_fail_closed_gpu_1929.py

Module helpers and fixtures that collided across sources are
suffixed with the source issue number so sibling sections stay
independent. GPU gating moves to the shared requires_gpu marker
from _helpers/markers.py (aliased _gpu_only).

Two pytest.importorskip calls moved from module scope to per-test
gating so unrelated sections still collect when one optional extra
(lerc, tifffile) is missing.

Docs pointers in docs/source/reference/geotiff.rst and
docs/source/reference/release_gate_geotiff.rst redirect to the new
module path so the release-gate checklist-parity test stays green.

163 tests collected (matching the pre-consolidation total).
