+++ xrspatial/geotiff/tests/gpu/__init__.py
@@ -1,0 +1,1 @@
+"""GPU-flavoured test cluster for the GeoTIFF module."""
