--- xrspatial/geotiff/tests/test_kvikio_batched_pread_1688.py
@@ -1,410 +1,0 @@
-"""Regression tests for the batched-pread + single-buffer pattern in
-_try_kvikio_read_tiles.
-
-Issue #1688: ``_try_kvikio_read_tiles`` used to allocate one ``cupy.empty(bc)``
-per tile and block on ``IOFuture.get()`` between successive ``pread`` calls.
-That forced the GDS reads to serialise in kvikio's worker pool and paid the
-per-tile cupy allocation cost N times. The fix:
-
-* pre-allocates one contiguous ``cupy.empty(sum(tile_byte_counts))`` buffer
-  guarded by ``_check_gpu_memory``;
-* submits every ``pread`` call before waiting on any of them;
-* returns ``list[cupy.ndarray]`` per-tile views into the shared buffer so
-  the downstream nvCOMP / batched-D2H consumers are unchanged.
-
-These tests skip when CuPy + CUDA are not available. The kvikio integration
-path uses a fake CuFile so the structural checks (single allocation,
-submit-then-wait ordering, memory guard) run on hosts without kvikio.
-"""
-from __future__ import annotations
-
-import importlib.util
-
-import numpy as np
-import pytest
-
-from xrspatial.geotiff._gpu_decode import _try_kvikio_read_tiles
-
-
-def _gpu_available() -> bool:
-    if importlib.util.find_spec("cupy") is None:
-        return False
-    try:
-        import cupy
-        return bool(cupy.cuda.is_available())
-    except Exception:
-        return False
-
-
-# ---------------------------------------------------------------------------
-# Fake kvikio CuFile used by tests that do not depend on a real GDS install.
-# Records pread submission order + offsets so the test can assert the new
-# "submit all, wait all" pattern.
-# ---------------------------------------------------------------------------
-
-
-class _FakeIOFuture:
-    """Stand-in for ``kvikio._lib.cufile.IOFuture``.
-
-    ``.get()`` returns the requested byte count; tests that exercise the
-    partial-read fallback construct one with ``bc - 1`` instead so the
-    function returns None.
-    """
-
-    def __init__(self, value):
-        self._value = value
-        self.get_called = False
-
-    def get(self):
-        self.get_called = True
-        return self._value
-
-
-class _RecordingCuFile:
-    """In-memory CuFile that records pread arguments and write order.
-
-    Writes deterministic bytes into the provided buffer so the test can
-    verify the result is a list of per-tile views over one contiguous
-    buffer (not a list of independent allocations).
-    """
-
-    def __init__(self, file_bytes):
-        self.file_bytes = file_bytes
-        self.preads = []  # list of (file_offset, length, buf_id)
-        self.pread_order = []  # order in which preads were submitted
-        self.gets_after_preads = 0
-        self._preads_seen = 0
-        self.closed = False
-
-    def __enter__(self):
-        return self
-
-    def __exit__(self, *exc):
-        self.closed = True
-        return False
-
-    def pread(self, buf, file_offset=0, size=None, task_size=None):
-        if size is None:
-            size = int(buf.size)
-        # Write requested file bytes into the device buffer for later
-        # round-trip verification. ``buf`` is a cupy view, so this is the
-        # H2D equivalent of the real GDS DMA.
-        import cupy
-
-        host_chunk = np.frombuffer(
-            self.file_bytes[file_offset:file_offset + size], dtype=np.uint8)
-        buf[:] = cupy.asarray(host_chunk)
-
-        # Capture submission order. Track the buffer's data pointer +
-        # offset so the test can assert all writes landed in one base
-        # allocation.
-        self.pread_order.append(
-            (int(file_offset), int(size), int(buf.data.ptr)))
-        self._preads_seen += 1
-        return _FakeIOFuture(size)
-
-
-class _PartialRecordingCuFile(_RecordingCuFile):
-    """CuFile whose nth pread short-reads (returns ``bc - 1`` bytes)."""
-
-    def __init__(self, file_bytes, fail_index):
-        super().__init__(file_bytes)
-        self.fail_index = fail_index
-
-    def pread(self, buf, file_offset=0, size=None, task_size=None):
-        if size is None:
-            size = int(buf.size)
-        future = super().pread(buf, file_offset=file_offset, size=size)
-        if len(self.pread_order) == self.fail_index + 1:
-            return _FakeIOFuture(size - 1)  # partial read
-        return future
-
-
-def _install_fake_kvikio(monkeypatch, cufile_cls):
-    """Make ``import kvikio`` inside ``_try_kvikio_read_tiles`` resolve to a
-    module whose ``CuFile`` is our recording stand-in.
-
-    The function imports kvikio lazily, so monkeypatching the module via
-    ``sys.modules`` is enough for both the install-present and
-    install-absent code paths to see the fake.
-    """
-    import sys
-    import types
-
-    fake_mod = types.ModuleType("kvikio")
-    fake_mod.CuFile = cufile_cls
-    monkeypatch.setitem(sys.modules, "kvikio", fake_mod)
-    return fake_mod
-
-
-# ---------------------------------------------------------------------------
-# Tests
-# ---------------------------------------------------------------------------
-
-
-def test_empty_tile_list_returns_empty_list():
-    """Zero tiles must return ``[]`` without touching cupy or kvikio."""
-    assert _try_kvikio_read_tiles("/nonexistent", [], [], 0) == []
-
-
-@pytest.mark.skipif(not _gpu_available(), reason="cupy + CUDA required")
-def test_kvikio_missing_returns_none(monkeypatch):
-    """When kvikio is not installed, the function must return None.
-
-    ``gpu_decode_tiles_from_file`` relies on this signal to switch to the
-    CPU mmap fallback. Without it, the caller would see an ImportError it
-    cannot recover from.
-    """
-    import sys
-
-    # Force the kvikio import inside _try_kvikio_read_tiles to ImportError.
-    monkeypatch.setitem(sys.modules, "kvikio", None)
-
-    result = _try_kvikio_read_tiles(
-        "/path/does/not/matter", [0, 1024], [1024, 1024], 1024)
-    assert result is None
-
-
-@pytest.mark.skipif(not _gpu_available(), reason="cupy + CUDA required")
-def test_single_buffer_allocation(monkeypatch):
-    """The fix allocates one contiguous device buffer, not N small ones.
-
-    Verified structurally: every pread's destination data-pointer must
-    fall within the single base allocation. The buffer pointer for tile i
-    must equal ``base_ptr + offset_i`` where ``offset_i = sum(sizes[:i])``.
-    """
-    import cupy
-
-    tile_offsets = [0, 4096, 8192, 12288]
-    tile_byte_counts = [1024, 2048, 512, 768]
-    file_size = max(o + bc for o, bc in zip(tile_offsets, tile_byte_counts))
-    file_bytes = np.arange(file_size, dtype=np.uint64).tobytes()[:file_size]
-
-    fake_cufile = _RecordingCuFile(file_bytes)
-    _install_fake_kvikio(monkeypatch, lambda path, mode='r': fake_cufile)
-
-    result = _try_kvikio_read_tiles(
-        "/fake/path.tif", tile_offsets, tile_byte_counts, max(tile_byte_counts))
-
-    assert result is not None
-    assert len(result) == len(tile_byte_counts)
-    for view, expected_bc in zip(result, tile_byte_counts):
-        assert isinstance(view, cupy.ndarray)
-        assert int(view.size) == expected_bc
-        assert view.dtype == cupy.uint8
-
-    # All views must share one base allocation: each view's pointer is
-    # ``base + sum(sizes[:i])``, with strictly monotonically increasing
-    # pointers separated by exactly the prior tile's size.
-    ptrs = [int(v.data.ptr) for v in result]
-    base = ptrs[0]
-    for i, view in enumerate(result):
-        expected = base + sum(tile_byte_counts[:i])
-        assert int(view.data.ptr) == expected, (
-            f"tile {i} pointer {int(view.data.ptr)} != expected {expected}; "
-            "per-tile allocations leaked back in?"
-        )
-
-
-@pytest.mark.skipif(not _gpu_available(), reason="cupy + CUDA required")
-def test_all_preads_submitted_before_any_get(monkeypatch):
-    """Every ``pread`` must be submitted before the first ``IOFuture.get()``.
-
-    The old loop alternated submit -> wait -> submit -> wait, which
-    serialised the IO in kvikio's worker pool. The new loop submits N
-    preads then waits on them in order; observing all N submissions
-    before any ``.get()`` is the structural signature of the fix.
-
-    The previous version of this test asserted on per-event lists
-    (``submission_log == [0,1,2,3]`` and ``get_log == [0,1,2,3]``)
-    which also passed under the legacy submit -> get -> submit -> get
-    loop. To actually catch a regression we record both events into a
-    single ordered timeline and assert every ``submit`` occurs before
-    the first ``get``.
-    """
-    events = []  # ('submit', tag) and ('get', tag) appended in real order
-
-    class _LoggingFuture(_FakeIOFuture):
-        def __init__(self, value, tag):
-            super().__init__(value)
-            self._tag = tag
-
-        def get(self):
-            events.append(('get', self._tag))
-            return super().get()
-
-    class _LoggingCuFile(_RecordingCuFile):
-        def pread(self, buf, file_offset=0, size=None, task_size=None):
-            if size is None:
-                size = int(buf.size)
-            tag = sum(1 for e in events if e[0] == 'submit')
-            events.append(('submit', tag))
-            super().pread(buf, file_offset=file_offset, size=size)
-            return _LoggingFuture(size, tag)
-
-    file_bytes = bytes(4096)
-    _install_fake_kvikio(
-        monkeypatch, lambda path, mode='r': _LoggingCuFile(file_bytes))
-
-    tile_offsets = [0, 256, 512, 768]
-    tile_byte_counts = [256, 256, 256, 256]
-    _try_kvikio_read_tiles(
-        "/fake/path.tif", tile_offsets, tile_byte_counts, 256)
-
-    n_tiles = len(tile_byte_counts)
-    submit_indices = [i for i, e in enumerate(events) if e[0] == 'submit']
-    get_indices = [i for i, e in enumerate(events) if e[0] == 'get']
-
-    # Sanity: every tile got submitted and waited on exactly once.
-    assert len(submit_indices) == n_tiles
-    assert len(get_indices) == n_tiles
-    assert [e[1] for e in events if e[0] == 'submit'] == list(range(n_tiles))
-    assert [e[1] for e in events if e[0] == 'get'] == list(range(n_tiles))
-
-    # The structural check that distinguishes batched from interleaved:
-    # every submission index must come before the first get index. The
-    # legacy submit -> get -> submit -> get loop interleaves these, so
-    # the first ``get`` lands at events[1] while the last ``submit``
-    # lands at events[6], failing this assertion.
-    assert max(submit_indices) < min(get_indices), (
-        "preads and gets are interleaved (legacy serial pattern); "
-        f"events timeline: {events}"
-    )
-
-
-@pytest.mark.skipif(not _gpu_available(), reason="cupy + CUDA required")
-def test_memory_guard_runs_with_total_byte_count(monkeypatch):
-    """The single-buffer allocation must be size-checked before ``cupy.empty``.
-
-    The OOM guard tells the caller early that the read will not fit on
-    the device. A regression that removed it would surface as an opaque
-    CUDA OOM only after the first ``cupy.empty`` failed.
-    """
-    from xrspatial.geotiff import _gpu_decode
-
-    seen = {"total_bytes": None, "what": None, "called": False}
-
-    def fake_check(required_bytes, what="tile buffer"):
-        seen["total_bytes"] = int(required_bytes)
-        seen["what"] = what
-        seen["called"] = True
-        raise MemoryError("simulated OOM")
-
-    file_bytes = bytes(4096)
-    _install_fake_kvikio(
-        monkeypatch, lambda path, mode='r': _RecordingCuFile(file_bytes))
-    monkeypatch.setattr(_gpu_decode, "_check_gpu_memory", fake_check)
-
-    tile_offsets = [0, 1024, 2048]
-    tile_byte_counts = [1024, 1024, 1024]
-
-    with pytest.raises(MemoryError, match="simulated OOM"):
-        _try_kvikio_read_tiles(
-            "/fake/path.tif", tile_offsets, tile_byte_counts, 1024)
-
-    assert seen["called"], "_check_gpu_memory was not called"
-    assert seen["total_bytes"] == sum(tile_byte_counts), (
-        f"expected total {sum(tile_byte_counts)}, got {seen['total_bytes']}"
-    )
-    assert "kvikio" in seen["what"] or "read buffer" in seen["what"], (
-        f"unhelpful 'what' label: {seen['what']!r}"
-    )
-
-
-@pytest.mark.skipif(not _gpu_available(), reason="cupy + CUDA required")
-def test_partial_read_returns_none(monkeypatch):
-    """When any pread reports fewer bytes than requested, the function
-    must return None so the caller falls back.
-    """
-    file_bytes = bytes(4096)
-    fail_at = 1  # second pread under-reads
-
-    def _factory(path, mode='r'):
-        return _PartialRecordingCuFile(file_bytes, fail_at)
-
-    _install_fake_kvikio(monkeypatch, _factory)
-
-    tile_offsets = [0, 256, 512]
-    tile_byte_counts = [256, 256, 256]
-    result = _try_kvikio_read_tiles(
-        "/fake/path.tif", tile_offsets, tile_byte_counts, 256)
-    assert result is None
-
-
-@pytest.mark.skipif(not _gpu_available(), reason="cupy + CUDA required")
-def test_round_trip_data_preserved(monkeypatch):
-    """End-to-end: the bytes read into the per-tile views must match the
-    bytes at the corresponding file offsets.
-
-    Distinct tiles get distinct payloads so a swap or off-by-one in the
-    offset bookkeeping would surface as a payload mismatch.
-    """
-    rng = np.random.default_rng(seed=1688)
-
-    tile_offsets = [0, 1024, 2048, 3072]
-    tile_byte_counts = [1024, 1024, 1024, 1024]
-    file_size = 4096
-    file_data = rng.integers(0, 256, size=file_size, dtype=np.uint8)
-    file_bytes = file_data.tobytes()
-
-    _install_fake_kvikio(
-        monkeypatch, lambda path, mode='r': _RecordingCuFile(file_bytes))
-
-    result = _try_kvikio_read_tiles(
-        "/fake/path.tif", tile_offsets, tile_byte_counts, 1024)
-
-    assert result is not None
-    assert len(result) == 4
-    for i, view in enumerate(result):
-        got = view.get()
-        want = file_data[tile_offsets[i]:tile_offsets[i] + tile_byte_counts[i]]
-        assert np.array_equal(got, want), f"tile {i} payload mismatch"
-
-
-@pytest.mark.skipif(not _gpu_available(), reason="cupy + CUDA required")
-def test_zero_size_tile_returns_zero_length_view(monkeypatch):
-    """A tile with ``byte_count == 0`` (sparse tile) must round-trip as a
-    zero-length view in the result list so the caller's iteration order
-    matches the original tile order.
-    """
-    import cupy
-
-    tile_offsets = [0, 1024, 1024, 2048]
-    tile_byte_counts = [1024, 0, 1024, 1024]
-    file_bytes = bytes(3072)
-
-    _install_fake_kvikio(
-        monkeypatch, lambda path, mode='r': _RecordingCuFile(file_bytes))
-
-    result = _try_kvikio_read_tiles(
-        "/fake/path.tif", tile_offsets, tile_byte_counts, 1024)
-
-    assert result is not None
-    assert len(result) == 4
-    assert int(result[1].size) == 0
-    assert isinstance(result[1], cupy.ndarray)
-    assert result[1].dtype == cupy.uint8
-
-
-@pytest.mark.skipif(not _gpu_available(), reason="cupy + CUDA required")
-def test_all_zero_size_tiles_returns_zero_length_views(monkeypatch):
-    """Edge: every tile is sparse (sum bytes == 0). Must return a list
-    of zero-length views without allocating a zero-sized buffer.
-    """
-    import cupy
-
-    # Note: this path does not hit kvikio at all (total_bytes == 0 short
-    # circuits before the CuFile is opened), so the kvikio module being
-    # absent is fine. Still install a fake to keep behaviour consistent
-    # if the total_bytes==0 short-circuit ever moves below the open().
-    _install_fake_kvikio(
-        monkeypatch, lambda path, mode='r': _RecordingCuFile(b""))
-
-    result = _try_kvikio_read_tiles(
-        "/fake/path.tif", [0, 0, 0], [0, 0, 0], 0)
-
-    assert result is not None
-    assert len(result) == 3
-    for view in result:
-        assert isinstance(view, cupy.ndarray)
-        assert int(view.size) == 0
